v1.1.0 for package rename, update to cdap6
