+++ README.md
@@ -67,7 +67,7 @@ CDAP Slack Channel: http://cdap-users.herokuapp.com/
 
 ## License and Trademarks
 
-Copyright © 2017 Cask Data, Inc.
+Copyright © 2017-2019 Cask Data, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
 in compliance with the License. You may obtain a copy of the License at
