+++ src/main/java/io/cdap/plugin/common/StreamCharSource.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2017 Cask Data, Inc.
+ * Copyright © 2017-2019 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,31 +14,26 @@
  * the License.
  */
 
-package co.cask.common;
+package io.cdap.plugin.common;
 
 import com.google.common.io.CharSource;
-import org.apache.twill.filesystem.Location;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
-import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 
 /**
- * A {@link CharSource} for {@link Location}.
+ * {@link CharSource} for the {@link InputStream}
  */
-public class LocationCharSource extends CharSource {
-
-  private final Location location;
-  private final Charset charset;
-
-  public LocationCharSource(Location location, Charset charset) {
-    this.location = location;
-    this.charset = charset;
+public class StreamCharSource extends CharSource {
+  private final InputStream is;
+  public StreamCharSource(InputStream is) {
+    this.is = is;
   }
-
   @Override
   public Reader openStream() throws IOException {
-    return new InputStreamReader(location.getInputStream(), charset);
+    return new InputStreamReader(is, StandardCharsets.UTF_8);
   }
 }
